logs1-orchestrator/requirements.txt
appdirs==1.4.4
attrs==21.2.0
Authlib==0.15.3
beautifulsoup4==4.9.3
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
comtypes==1.1.7
cryptography==3.4.7
decorator==5.0.7
defusedxml==0.7.1
dnspython==2.1.0
docutils==0.17.1
et-xmlfile==1.1.0
exchangelib==3.3.2
fpdf2==2.3.4
furl==2.1.2
graphviz==0.13.2
idna==2.10
isodate==0.6.0
jsonpath-ng==1.5.2
jsonschema==3.2.0
loguru==0.5.3
lxml==4.6.3
mss==6.1.0
netsuitesdk==1.23.0
notifiers==1.2.1
ntlm-auth==1.5.0
oauthlib==3.1.0
openpyxl==3.0.7
orderedmultidict==1.0.1
pdfminer.six==20201018
Pillow==8.2.0
ply==3.11
psutil==5.8.0
pycparser==2.20
Pygments==2.9.0
pynput-robocorp-fork==2.0.0
PyPDF2==1.26.0
pyperclip==1.8.2
pyrsistent==0.17.3
PySocks==1.7.1
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-ntlm==1.1.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3987==1.3.8
robotframework==4.0.2
robotframework-pythonlibcore==2.2.1
robotframework-requests==0.7.2
robotframework-seleniumlibrary==5.1.3
robotframework-seleniumtestability==1.1.0
rpaframework==9.6.0
rpaframework-core==6.2.0
rpaframework-pdf==0.5.1
selenium==3.141.0
simple-salesforce==1.11.1
six==1.16.0
sortedcontainers==2.3.0
soupsieve==2.2.1
tqdm==4.60.0
tweepy==3.10.0
tzdata==2021.1
tzlocal==2.1
urllib3==1.26.4
webdrivermanager==0.10.0
wrapt==1.12.1
xlrd==2.0.1
xlutils==2.0.0
xlwt==1.3.0
zeep==4.0.0
